// interfaces/interface_Cards.ts
import {Medium} from "@/interfaces/interface_globals"

export interface PistolCard {
    mainCategory: string
    subCategory: string
    country: string
    manufacturer: string
    name: string
    image: Medium
    length: string
    width: string
    height: string
    triggerPull: string
    remarks: string
    _modified: number
    _mby: string
    _created: number
    _state: number
    _cby: string
    _id: string
  }
  